+++ docs/.vitepress/config.ts
@@ -23,7 +23,8 @@ export default defineConfig({
       {
         text: 'Versions',
         items: [
-          { text: '3.x.x', link: '/guide/essentials/introduction' },
+          { text: '4.x.x', link: '/guide/essentials/introduction' },
+          { text: '3.x.x', link: '/v3/guide/essentials/introduction' },
           { text: '2.4.2', link: '/v2/guide/essentials/introduction' },
         ]
 
@@ -110,6 +111,10 @@ export default defineConfig({
           {
             text: 'Advanced Usage',
             items: [
+              {
+                text: 'Regeneration of variants',
+                link: '/guide/advanced_usage/regenerate-variant',
+              },
               {
                 text: 'Exceptions',
                 link: '/guide/advanced_usage/exceptions',
@@ -253,8 +258,113 @@ export default defineConfig({
     {
       text: 'ChangeLog',
       link: '/changelog',
-    },
-  ],
+    }],
+  '/v3/': [{
+    text: 'Guide',
+    items: [
+      {
+        text: 'Start here',
+        link: '/v3/guide/start-here',
+      },
+      {
+        text: 'Essentials',
+        items: [
+          {
+            text: 'Introduction',
+            link: '/v3/guide/essentials/introduction',
+          },
+          {
+            text: 'Installation',
+            link: '/v3/guide/essentials/installation',
+          },
+          {
+            text: 'Configuration',
+            link: '/v3/guide/essentials/configuration',
+          },
+        ],
+      },
+      {
+        text: 'Basic Usage',
+        items: [
+          {
+            text: 'Migration Setup',
+            link: '/v3/guide/basic_usage/migration-setup',
+          },
+          {
+            text: 'Model Setup',
+            link: '/v3/guide/basic_usage/model-setup',
+          },
+          {
+            text: 'Controller Setup',
+            link: '/v3/guide/basic_usage/controller-setup',
+          },
+          {
+            text: 'View Setup',
+            link: '/v3/guide/basic_usage/view-setup',
+          },
+        ],
+      },
+      {
+        text: 'Converter',
+        items: [
+          {
+            text: 'Image',
+            link: '/v3/guide/converters/image',
+          },
+          {
+            text: 'PDF thumbnail',
+            link: '/v3/guide/converters/pdf-thumbnail',
+          },
+          {
+            text: 'Document thumbnail',
+            link: '/v3/guide/converters/document-thumbnail',
+          },
+          {
+            text: 'Video thumbnail',
+            link: '/v3/guide/converters/video-thumbnail',
+          },
+        ],
+      },
+      {
+        text: 'Advanced Usage',
+        items: [
+          {
+            text: 'Exceptions',
+            link: '/v3/guide/advanced_usage/exceptions',
+          },
+          {
+            text: 'PrecompileUrl',
+            link: '/v3/guide/advanced_usage/pre-compile-on-demand',
+          },
+          {
+            text: 'Custom converter',
+            link: '/v3/guide/advanced_usage/custom-converter',
+          },
+          {
+            text: 'Queue',
+            link: '/v3/guide/advanced_usage/queue',
+          },
+        ],
+      },
+      {
+        text: 'Use cases',
+        items: [
+          {
+            text: 'Picture',
+            link: '/v3/guide/use-cases/picture',
+          },
+        ],
+      },
+    ],
+  },
+  {
+    text: 'Structure data JSON',
+    link: '/structure-data-json',
+  },
+  {
+    text: 'ChangeLog',
+    link: '/changelog',
+  }],
   },
 
     socialLinks: [
@@ -270,6 +380,7 @@ export default defineConfig({
           const html = md.render(src, env)
           if (env.frontmatter?.search === false) return ''
           if (env.relativePath.startsWith('v2/')) return ''
+          if (env.relativePath.startsWith('v3/')) return ''
           return html
         }
       }